Renders the Chart locally, using the 'helm template ...' command
This is possible since this is a Helm Plugin,
so the Helm command should always be available.
However, a better strategy could have been
to leverage the Helm Go SDK right away,
but it looks like the logic behind the
'helm template' command cannot be easily called from the outside.